rust-server-cleanup(config-module,non_mechanical_provable): extract launch configuration helpers
